requirements.txt
selenium==4.16.0
pymongo==4.6.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
websockets==12.0
pydantic==2.5.2
loguru==0.7.2